Add all sorts to investigation table
